+++ server/operator/operator.go
@@ -2,7 +2,6 @@ package operator
 
 import (
 	"github.com/phzfi/RIC/server/cache"
-	"github.com/phzfi/RIC/server/images"
 	"github.com/phzfi/RIC/server/ops"
 	"sync"
 )
@@ -12,7 +11,8 @@ type Operator struct {
 
 	sync.Mutex
 	inProgress map[string]*Progress
-	tokens     TokenPool
+
+	processor imageProcessor
 }
 
 type Progress struct {
@@ -29,7 +29,7 @@ func Make(cache Cacher) Operator {
 	return Operator{
 		cache:      cache,
 		inProgress: make(map[string]*Progress),
-		tokens:     MakeTokenPool(2),
+		processor:  makeImageProcessor(2),
 	}
 }
 
@@ -76,7 +76,7 @@ func (o *Operator) GetBlob(operations ...ops.Operation) (blob []byte, err error)
 		var found bool
 		blob, found = o.cache.GetBlob(key)
 		if !found {
-			blob, err = o.makeBlob(startimage, operations[start:])
+			blob, err = o.processor.makeBlob(startimage, operations[start:])
 			if err != nil {
 				return nil, err
 			}
@@ -101,25 +101,4 @@ func (o *Operator) addInProgress(key string) *Progress {
 
 	o.inProgress[key] = p
 	return p
-}
-
-func (o *Operator) makeBlob(startBlob []byte, operations []ops.Operation) ([]byte, error) {
-	o.tokens.Borrow()
-	defer o.tokens.Return()
-
-	img := images.NewImage()
-	defer img.Destroy()
-
-	if startBlob != nil {
-		img.FromBlob(startBlob)
-	}
-
-	for _, op := range operations {
-		err := op.Apply(img)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return img.Blob(), nil
 }
